add temporary methods

+++ Plugin~/WebRTCPlugin/Codec/NvCodec/NvEncoder.cpp
@@ -235,7 +235,7 @@ namespace webrtc
     {
         const uint32_t bitrate = parameters.bitrate.get_sum_bps();
         m_bitrateAdjuster->SetTargetBitrateBps(bitrate);
-        m_frameRate = parameters.framerate_fps;
+        m_frameRate = static_cast<uint32_t>(parameters.framerate_fps);
     }
 
     bool NvEncoder::CopyBuffer(void* frame)
